Re-organized class objects and refined documentation
