+++ BridgeApp/BridgeApp/Data Tracking/iOS/SBARemoveTrackedItemStepViewController.swift
@@ -77,77 +77,3 @@ open class SBARemoveTrackedItemStepViewController: RSDStepViewController {
         self.textLabel?.text = self.removeTrackedItemStep?.text
     }
 }
-
-/// `SBARemoveTrackedItemStepObject` is a simple instruction step that includes a title
-/// and the option to underline a text segment of the title.
-open class SBARemoveTrackedItemStepObject: RSDUIStepObject, RSDStepViewControllerVendor {
-    
-    public func instantiateViewController(with taskPath: RSDTaskPath) -> (UIViewController & RSDStepController)? {
-        let vc = SBARemoveTrackedItemStepViewController(nibName: SBARemoveTrackedItemStepViewController.nibName, bundle: SBARemoveTrackedItemStepViewController.bundle)
-        vc.step = self
-        return vc
-    }
-    
-    private enum CodingKeys: String, CodingKey {
-        case bodyText, underlinedTitleSegment, items
-    }
-    
-    /// The phrase or text segment that will be underlined in the title
-    public var underlinedTitleSegment: String?
-    
-    /// The list of items to be removed
-    public var items: [RSDIdentifier]
-    
-    override open func instantiateStepResult() -> RSDResult {
-        return SBARemoveTrackedItemsResultObject(identifier: self.identifier, items: self.items)
-    }
-    
-    public init(identifier: String, title: String, underlinedTitleSegment: String, items: [RSDIdentifier]) {
-        self.items = items
-        super.init(identifier: identifier, type: .removeTrackedItem)
-        self.underlinedTitleSegment = underlinedTitleSegment
-        self.title = title
-    }
-    
-    /// Initializer required for `copy(with:)` implementation.
-    public required init(identifier: String, type: RSDStepType?) {
-        self.items = [RSDIdentifier(rawValue: identifier)]
-        super.init(identifier: identifier, type: type ?? .removeTrackedItem)
-    }
-    
-    /// Override to set the properties of the subclass.
-    override open func copyInto(_ copy: RSDUIStepObject) {
-        super.copyInto(copy)
-        guard let subclassCopy = copy as? SBARemoveTrackedItemStepObject else {
-            assertionFailure("Superclass implementation of the `copy(with:)` protocol should return an instance of this class.")
-            return
-        }
-        subclassCopy.underlinedTitleSegment = self.underlinedTitleSegment
-        subclassCopy.items = self.items
-    }
-    
-    /// Initialize from a `Decoder`.
-    ///
-    /// - example:
-    /// ```
-    ///    let json = """
-    ///        {
-    ///            "identifier": "foo",
-    ///            "type": "selection",
-    ///            "title": "Please select the items you wish to remove",
-    ///            "underlinedTitleSegment": "the items",
-    ///            "items": ["itemA", "itemB"]
-    ///        }
-    ///        """.data(using: .utf8)! // our data in native (JSON) format
-    /// ```
-    ///
-    /// - parameter decoder: The decoder to use to decode this instance.
-    /// - throws: `DecodingError`
-    public required init(from decoder: Decoder) throws {
-        // Decode the body text and underlined body text segment.
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.underlinedTitleSegment = try container.decodeIfPresent(String.self, forKey: .underlinedTitleSegment)
-        self.items = try container.decode([RSDIdentifier].self, forKey: .items)
-        try super.init(from: decoder)
-    }
-}
